CoreJava_yk/src/com/crmindz/bcjmay2017/Collections/testcases/ReverseStringsTest.java
package com.crmindz.bcjmay2017.Collections.testcases;


import static org.junit.Assert.assertEquals;

import java.util.ArrayList;

import org.junit.Test;

import com.crmindz.bcjmay2017.Collections.lists.lab1.ReverseStrings;

/**
 * @author Yeshwanth Konakanchi
 * This is test class for the Reverse Strings class
 */
public class ReverseStringsTest {

	@Test
	public void testReverseStrings() {
		
		ArrayList<String> input = new ArrayList<String>();
		
		input.add("Siva");
		input.add("Yarram");
		
		ArrayList<String> expected = new ArrayList<String>();
		
		expected.add("aviS");
		expected.add("marraY");
		
		assertEquals(expected,ReverseStrings.reverseStrings(input));

		
	}
	
	@Test
	public void testReverseStrings2() {
		
		ArrayList<String> input = new ArrayList<String>();
		
		input.add("Sravya");
		input.add("Yeleti");
		
		ArrayList<String> expected = new ArrayList<String>();
		
		expected.add("ayvarS");
		expected.add("iteleY");
		
		assertEquals(expected,ReverseStrings.reverseStrings(input));

		
	}


}
